% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk.R
\name{sk}
\alias{sk}
\title{the sample skewness coefficient}
\usage{
sk(x)
}
\arguments{
\item{x}{the data which is a numeric vector}
}
\value{
A numeric vector of length one
}
\description{
This function is used to compute the skewness coefficient of a set of numeric data x
}
\examples{
\dontrun{
x <- rnorm(100)
sk(x)
}
}
